real-time-polling-app/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.30
mysql-connector-python==8.4.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
websockets==12.0